+++ Projetos/Projeto Final/src/visual/ListaPorPefil.java
@@ -324,7 +324,7 @@ public class ListaPorPefil extends JPanel {
 		panel.add(cepDoCliente);
 		cepDoCliente.setFont(new Font("Californian FB", Font.BOLD | Font.ITALIC, 18));
 		
-		JLabel cpfDoCliente = new JLabel(Integer.toString(livraria.getClientes().get(lista.getSelectedIndex()).getCPF()));
+		JLabel cpfDoCliente = new JLabel(Long.toString(livraria.getClientes().get(lista.getSelectedIndex()).getCPF()));
 		cpfDoCliente.setBounds(77, 231, 417, 19);
 		panel.add(cpfDoCliente);
 		cpfDoCliente.setFont(new Font("Californian FB", Font.BOLD | Font.ITALIC, 18));
